% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/philtools.R
\name{\%after\%}
\alias{\%after\%}
\alias{\%before\%}
\title{Comparing POSIXct objects}
\usage{
ct \%after\% st

ct \%before\% st
}
\arguments{
\item{ct}{Value to be compared}

\item{st}{Value to be compared}
}
\value{
TRUE or FALSE, the result of the comparison
}
\description{
Use %before% and %after% as infix operators between POSIXct objects and strings.
}
\examples{
 as.POSIXct("2015-01-01") \%before\% "2015-01-02"

}

