% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/philtools.R
\name{\%~\%}
\alias{\%~\%}
\title{Regular Expression Matching}
\usage{
string \%~\% pattern
}
\arguments{
\item{string}{The string to be searched}

\item{regex}{The regex to be matched}
}
\value{
TRUE or FALSE, the result of the regex search
}
\description{
Use %~% to run a regex against a string, like perl.
}
\examples{
"banana" \%~\% "(an){2}"

}

