% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/philtools.R
\name{scale_y_digital}
\alias{scale_x_digital}
\alias{scale_y_digital}
\title{Position scales for digital information (x & y)}
\usage{
scale_y_digital(nbreaks = 4, ...)

scale_x_digital(nbreaks = 4, ...)
}
\arguments{
\item{nbreaks}{Integer count of breaks required}

\item{...}{Other arguments passed on to scale_(x|y)_continuous}
}
\description{
Continuous scales with breaks set to round numbers of powers of two,
and labels using K, M, G, T, P etc.
}
\examples{
diamonds \%>\% qplot(data=., x=carat, y=price) + scale_y_digital()

}

